requirements.txt
pandas==1.5.1
plotly==5.11.0
psycopg2-binary==2.9.5
pyTelegramBotAPI==4.7.1
requests==2.28.1

